requirements.txt
python-telegram-bot==20.0
spleeter==2.0.0
noisereduce==3.0.0
numpy==1.24.3
pydub==0.25.1
rarfile==4.0


